+++ tests/testthat/test-free_cores.R
@@ -1,0 +1,7 @@
+test_that("free_cores throws deprecation warning", {
+  skip_on_cran()
+  expect_warning(
+    free_cores(),
+    regexp = "`free_cores\\(\\)` was deprecated in mmrm 0\\.1\\.6."
+  )
+})